Inject TimeProvider into Android sync outbox and clean up touched code
Rename ProgressTimeProvider/SystemProgressTimeProvider to the generic
TimeProvider/SystemTimeProvider since the abstraction is now used outside
Progress, and split the Progress-specific snapshot helpers into their own
file. Inject the new TimeProvider into SyncOutboxLocalStore so outbox
createdAt is no longer pinned to System.currentTimeMillis directly. Add
an explanatory SQL comment to the migration 15->16 backfill, and drop the
unused progress_reviews_subtitle string resource.
