+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/TimeProvider.kt
@@ -1,0 +1,18 @@
+package com.flashcardsopensourceapp.data.local.repository
+
+import java.time.ZoneId
+
+interface TimeProvider {
+    fun currentZoneId(): ZoneId
+    fun currentTimeMillis(): Long
+}
+
+object SystemTimeProvider : TimeProvider {
+    override fun currentZoneId(): ZoneId {
+        return ZoneId.systemDefault()
+    }
+
+    override fun currentTimeMillis(): Long {
+        return System.currentTimeMillis()
+    }
+}
