+++ apps/android/feature/progress/src/main/res/values/strings.xml
@@ -13,7 +13,6 @@
     <string name="progress_streak_summary_reviewed_today">Reviewed today</string>
     <string name="progress_streak_summary_not_reviewed_today">Not reviewed today</string>
     <string name="progress_reviews_title">Reviews</string>
-    <string name="progress_reviews_subtitle">Last 20 weeks</string>
     <string name="progress_reviews_previous_week">Previous week</string>
     <string name="progress_reviews_next_week">Next week</string>
     <string name="progress_reviews_empty_week">No reviews yet in this week.</string>
